ai_service/requirements.txt
pydantic
google-cloud-storage
transformers
psycopg2-binary
grpcio
grpcio-tools
google-generativeai
google-genai
ultralytics
PyMuPDF
Pillow
torch==2.7.0 # Hoặc phiên bản cụ thể 